Tweak fixture semantics
The fixture setup code doesn't have convenient access to config file
settings, and also keying the automagic-async-fixture-handling off of
the trio-mode setting would be a compatibility break. Not the end of
the world, but kind of annoying.

Instead, let's continue to key the automagic-async-fixture-handling
off of the requesting test being a Trio test. It's just as convenient,
and while technically could be a little more error-prone –
specifically, if you accidentally use an async fixture with a
non-async test, then we won't be able to catch that as an error – I
doubt it matters much in practice. And if people turn out to make that
mistake a lot, we can always add some check for it later.