Fix compatibility with pytest 3.8.1
Apparently for a test *method*, item.obj and item.function are
different: item.obj is the bound method, and item.function is the
underlying unbound method. At least in pytest 3.8.1.

We want to wrap the bound method, not the unbound method, so we should
wrap item.obj, not item.function.